requirements.txt
requests~=2.27.1
beautifulsoup4~=4.10.0
PyTelegramBotAPI
Pillow~=9.0.1